check for category existence for product action methods

+++ Controllers/ProductController.cs
@@ -1,8 +1,6 @@
 using api.DTOs;
 using api.Interfaces;
 using api.Mappers;
-using api.Models;
-using api.Repository;
 using Microsoft.AspNetCore.Mvc;
 
 namespace api.Controllers;
@@ -19,46 +17,52 @@ public class ProductController : ControllerBase
         _categoryRepo = categoryRepository;
         _productRepo = productRepository;
     }
-    
+
     [HttpGet]
-    public async Task<IEnumerable<ProductDto>> GetProductsByCategory(Guid categoryId)
+    public async Task<IEnumerable<ProductDto>> Index(Guid categoryId)
     {
         return await _productRepo.GetProductsByCategoryIdAsync(categoryId);
     }
 
     [HttpGet("{productId:guid}")]
-    public async Task<IActionResult> Get(Guid categoryId, Guid productId)
+    public async Task<IActionResult> Show(Guid categoryId, Guid productId)
     {
-        var product = await _productRepo.GetProductByIdAsync(productId);
-        if (product == null)
-        {
-            return NotFound($"Product with Id '{productId}' not found.");
-        }
-
-        return Ok(product);
-    }
-
-    [HttpPost]
-    public async Task<IActionResult> Create(Guid categoryId, CreateProductRequestDto productRequestDto)
-    {
-        if (!ModelState.IsValid)
-        {
-            return BadRequest(ModelState);
-        }
-        
         var categoryExists = await _categoryRepo.CategoryExists(categoryId);
         if (!categoryExists)
         {
             return NotFound($"Category with id '{categoryId}' not found.");
         }
-        
+
+        var productDto = await _productRepo.GetProductByIdAsync(productId);
+        if (productDto == null)
+        {
+            return NotFound($"Product with Id '{productId}' not found.");
+        }
+
+        return Ok(productDto);
+    }
+
+    [HttpPost]
+    public async Task<IActionResult> Store(Guid categoryId, CreateProductRequestDto productRequestDto)
+    {
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
+        }
+
+        var categoryExists = await _categoryRepo.CategoryExists(categoryId);
+        if (!categoryExists)
+        {
+            return NotFound($"Category with id '{categoryId}' not found.");
+        }
+
         var product = productRequestDto.ProductFromCreateRequestDto(categoryId);
-        
+
         var createdProduct = await _productRepo.CreateAsync(product);
 
-        return CreatedAtAction(nameof(Get), new { categoryId = categoryId, productId = product.Id }, createdProduct);
+        return CreatedAtAction(nameof(Show), new { categoryId, productId = product.Id }, createdProduct);
     }
-    
+
     [HttpPut("{productId:guid}")]
     public async Task<IActionResult> Update(Guid categoryId, Guid productId, UpdateProductRequestDto productRequestDto)
     {
@@ -66,13 +70,13 @@ public class ProductController : ControllerBase
         {
             return BadRequest(ModelState);
         }
-        
+
         var categoryExists = await _categoryRepo.CategoryExists(categoryId);
         if (!categoryExists)
         {
             return NotFound($"Category with id '{categoryId}' not found.");
         }
-        
+
         var product = productRequestDto.ToProductFromUpdateDto(categoryId);
 
         var updatedProduct = await _productRepo.UpdateAsync(productId, product);
@@ -85,10 +89,16 @@ public class ProductController : ControllerBase
     }
 
     [HttpDelete("{productId:guid}")]
-    public async Task<IActionResult> Delete(Guid categoryId, Guid productId)
+    public async Task<IActionResult> Destroy(Guid categoryId, Guid productId)
     {
-        var product = await _productRepo.DeleteAsync(productId);
-        if (product == null)
+        var categoryExists = await _categoryRepo.CategoryExists(categoryId);
+        if (!categoryExists)
+        {
+            return NotFound($"Category with id '{categoryId}' not found.");
+        }
+
+        var productDeleted = await _productRepo.DeleteAsync(productId);
+        if (productDeleted == false)
         {
             return NotFound($"Product with ID '{productId}' not found.");
         }
